Move stub/parser/types.rs items to stub.rs
